f32ext: tan(x)
Implements tangent in terms of `sin` and `cos`.

Note that `sin` is implemented in terms of `cos`, so this ultimately
implements it in terms of the cosine function alone.

In order to get decent precision (`~0.6` radians, which is difficult for
larger tangent values) this commit includes modifications to `cos` to
increase the precision of its approximation.