Refactor vector types (#69)
This is a fairly large SemVer-breaking overhaul of the vector types
provided by this crate which still largel manages to preserve the
original API while dramatically simplifying the implementation.

It removes all usage of `generic-array`. The original intention starting
this refactor was to leverage `min_const_generics` in its place, but it
turns out that const generics really aren't necessary and the previous
usage of `GenericArray` was largely superfluous.

It replaces all previous use of macros with `Vector2d`/`Vector3d` types
which are generic over coordinates. The previous concrete vector structs
such as `I32x2`, `I32x3`, `F32x2`, `F32x3` etc are now implemented as
type aliases for the generic structs.

It also eliminates the `VectorExt` trait as what it provided can be
implemented in terms of trait bounds on the individual methods.