Add `F32` newtype
Adds a newtype for `f32` which can be used as an alternative to
`F32Ext`.

The longer-term goal is to implement all of the current floating point
operations provided by this library on the newtype, and use the newtype
to impl `F32Ext`.

This will also allow us to implement traits from the `num-traits` crate
(#57).